Ask the file system for files instead of for handles
Everything that touched a file went through CreateFileW, GetFileAttributesW,
MoveFileExW, FindFirstFileW, SHGetKnownFolderPath and SearchPathW, scattered
across config, the screenshot path, the locator, the remuxer and the ffmpeg
download. Each of them had its own small copy of "is that file there" and its
own idea of what a folder is called.

files.h is the seam, and it is deliberately split in two. files.cpp holds what
the standard library answers the same way everywhere: exists, is-a-file,
list-a-folder, remove, rename, the temp folder. files_win32.cpp holds the rest,
which is the part std cannot promise -- a whole-file read that lets another
program keep the file open, a replace that is either the old file or the new one
and never half of each, the user's Videos and Pictures folders, a temp file that
is marked temporary, and the PATH search that turns "ffmpeg" into a path. A
second backend copies none of the first half.

archive.h is separate from it for one reason: only the two callers that download
something need a SHA-256 and an unzip, and keeping them apart is why the probe
links neither bcrypt nor the tar path. The implementation is the code that was
inside ffmpeg_download, moved out unchanged -- same 1 MB hashing buffer, same
tar.exe out of System32, same argument list.

std::filesystem is treated as the interface here, not as an implementation
detail to be hidden: paths stay paths, and PathToUtf8 is only reached for when
something has to be said to a process or written into a log. That took the last
wstring out of the locator, the remuxer and the screenshot naming.

Deliberate divergences, all of them narrowing behaviour rather than widening it:
ForeignSettings::modified is now seconds since 1970 instead of a Windows file
time, which sorts and displays the same and no longer means anything only
Windows can read; the ffmpeg target folder is created with its parents rather
than one level; "is that still there" became is-a-file where a folder of that
name used to count; removing a file that was already gone counts as success; and
the settings file this build writes is told apart from the others by an
ASCII-only case fold instead of _wcsicmp, which is the same answer for every
name the program has ever had.

Measured: pictures identical for all eight capture cases, frame age 0.9 to
1.7 ms against 1.0 to 1.4 ms before, drops equal or lower. Screenshots written
to a two-level folder that did not exist yet, in PNG and JPEG, with the
collision suffix appearing for three shots inside the same second. Two
recordings to AV1 with ffmpeg found through the configured path, both exit code
0. The remuxer and the AVIF still picture were read and not run.

One leftover found on the way: recorder.cpp still had a ::Sleep(2) between two
calls that had already been converted.

+++ src/config.h
@@ -771,10 +771,10 @@ struct ForeignSettings {
   // profiles back.
   int language = -1;
 
-  // Last written, as a Windows file time. Shown in the question, because when
-  // each file was last touched is the one fact that actually tells the two
-  // apart.
-  unsigned long long modified = 0;
+  // Last written, in seconds since 1970-01-01 UTC, 0 when unknown. Shown in
+  // the question, because when each file was last touched is the one fact that
+  // actually tells the two apart.
+  int64_t modified = 0;
 };
 
 // Every settings file next to the executable that is ours but is not the one
